+++ src/Ui/Bundle/EventBundle/Controller/Contact/ShowAction.php
@@ -1,0 +1,84 @@
+<?php
+
+/*
+ * This file is part of the Proximum Vimeet project.
+ *
+ * Copyright (C) Proximum
+ *
+ * @author Elao <contact@elao.com>
+ */
+
+namespace Proximum\Vimeet\Ui\Bundle\EventBundle\Controller\Contact;
+
+use Proximum\Vimeet\Application\Adapter\AuthorizationCheckerAdapterInterface;
+use Proximum\Vimeet\Application\Adapter\QueryBusInterface;
+use Proximum\Vimeet\Application\Query\Contact\GetContactViewQuery;
+use Proximum\Vimeet\Domain\Model\Sheet;
+use Proximum\Vimeet\Domain\Model\User;
+use Proximum\Vimeet\Domain\User\Sheet\HasAccessToSheet;
+use Proximum\Vimeet\Ui\Bundle\EventBundle\ParamConverter\EventDomain;
+use Proximum\Vimeet\Ui\Bundle\EventBundle\Security\SheetVoter;
+use Proximum\Vimeet\Ui\Bundle\EventBundle\ValueResolver\UserDomain;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+use Symfony\Component\Templating\EngineInterface;
+
+class ShowAction
+{
+    /** @var AuthorizationCheckerAdapterInterface */
+    private $authorizationChecker;
+
+    /** @var EngineInterface */
+    private $engine;
+
+    /** @var HasAccessToSheet */
+    private $hasAccessToSheet;
+
+    /** @var QueryBusInterface */
+    private $queryBus;
+
+    public function __construct(
+        AuthorizationCheckerAdapterInterface $authorizationChecker,
+        EngineInterface $engine,
+        HasAccessToSheet $hasAccessToSheet,
+        QueryBusInterface $queryBus
+    ) {
+        $this->authorizationChecker = $authorizationChecker;
+        $this->engine = $engine;
+        $this->hasAccessToSheet = $hasAccessToSheet;
+        $this->queryBus = $queryBus;
+    }
+
+    public function __invoke(
+        Request $request,
+        EventDomain $eventDomain,
+        UserDomain $userDomain,
+        Sheet $sheet,
+        User $contact
+    ): Response {
+        $event = $eventDomain->getEvent();
+        $user = $userDomain->getUser();
+
+        if (!$this->authorizationChecker->isGranted(SheetVoter::EDIT, $sheet)
+            || !$this->hasAccessToSheet->isSatisfiedBy($user, $event, $sheet)
+        ) {
+            throw new AccessDeniedException();
+        }
+
+        $contactView = $this->queryBus->handle(
+            new GetContactViewQuery($event, $sheet, $contact, $request->getLocale())
+        );
+
+        return new Response(
+            $this->engine->render(
+                '@Event/Contact/show.html.twig',
+                [
+                    'event' => $event,
+                    'sheet' => $sheet,
+                    'contactView' => $contactView,
+                ]
+            )
+        );
+    }
+}